fix: use int64 to prevent type change across architectures (Fixes #80)
Previously, the code relied on the default int type, which can vary
between 32-bit and 64-bit systems. This caused inconsistent behavior
and potential overflow issues. Using int64 ensures consistent
behavior across platforms.

+++ opnsense/unbound_dns.go
@@ -192,12 +192,12 @@ type unboundDNSStatusResponse struct {
 type UnboundDNSOverview struct {
 	AnswerRcodes      map[string]int
 	QueryTypes        map[string]int
-	Total             int
-	BlocklistSize     int
-	Passed            int
-	AnswerRcodesTotal int
-	AnnswerBogusTotal int
-	AnswerSecureTotal int
+	Total             int64
+	BlocklistSize     int64
+	Passed            int64
+	AnswerRcodesTotal int64
+	AnnswerBogusTotal int64
+	AnswerSecureTotal int64
 	UptimeSeconds     float64
 }
